refactor: rename pasta to config
